small refactor on settings workflow history container (#274)
