+++ client/containers/settings-workflow-history/get-default-state.js
@@ -19,12 +19,9 @@
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 // THE SOFTWARE.
 
-import Component from './component';
-import Connector from './connector';
-import getDefaultState from './get-default-state';
-import getters from './getters';
-import mutations from './mutations';
+const getDefaultState = (state = {}) => ({
+  graphEnabled: true,
+  ...state,
+});
 
-const container = Connector('SettingsWorkflowHistory', Component);
-
-export { container, getDefaultState, getters, mutations };
+export default getDefaultState;
